+++ Assets/Samples/Youtube Player/3.2.0/VideoInfo/PrintVideoInfo.cs
@@ -1,4 +1,6 @@
+using System.Linq;
 using UnityEngine;
+using UnityEngine.UI;
 using YoutubePlayer.Components;
 using YoutubePlayer.Extensions;
 
@@ -9,6 +11,8 @@ namespace YoutubePlayer.Samples.VideoInfo
         public InvidiousInstance InvidiousInstance;
 
         public string VideoId;
+
+        public Image Thumbnail;
 
         async void Start()
         {
@@ -23,6 +27,26 @@ namespace YoutubePlayer.Samples.VideoInfo
             infoString += $"Length: {videoInfo.LengthSeconds} seconds\n";
 
             Debug.Log(infoString);
+
+            Debug.Log("Loading thumbnail into texture");
+
+            var thumbnail = videoInfo.VideoThumbnails.First(thumbnail => thumbnail.Quality == "medium");
+            var texture = await thumbnail.ToTextureAsync();
+            Thumbnail.sprite = Sprite.Create(texture, new Rect(0, 0, texture.width, texture.height), new Vector2(0.5f, 0.5f));
+            Thumbnail.preserveAspect = true;
+
+            Debug.Log("Thumbnail loaded");
+
+            Debug.Log("You can also access the thumbnail using: " + InvidiousInstance.GetVideoThumbnailUrl(VideoId));
+        }
+
+        void OnDestroy()
+        {
+            if (Thumbnail != null && Thumbnail.sprite != null)
+            {
+                Destroy(Thumbnail.sprite.texture);
+                Destroy(Thumbnail.sprite);
+            }
         }
     }
 }