+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Loading thumbnails into textures
+
 ## [3.2.0] - 2024-04-13
 
 - 360 sample playing the video with the equi-rectengular projection format. No audio.